Update pom.xml
Switched from Eclipse's weird two-space indenting to standard four-space indenting, and added Javadocs and Sources as JAR goals, long with HTML-encoded versions of the source code that are accessible from the Javadocs (as this is open source code). The next step is to figure out how to publish this to Maven Central.