Update pom.xml
Even though the encoding for Maven takes the Properties value, I will now mimic JFreeOrg's redundant specification, which simply refers to the symbol defined in the properties, as it may be a failsafe in case something resets the encoding somewhere else (perhaps this is why the Eclipse Maven plug-in wasn't honouring UTF-8, meaning it's now safe for me to go back to its wizard vs. manually editing the POM files outside of Eclipse?).